+++ api/src/main/java/com/fernando/Repositories/BankRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fernando.Entities.Bank;
 
-public interface BankRepository extends JpaRepository<Bank, Integer> {
+public interface BankRepository extends JpaRepository<Bank, Long> {
 	
 }
